Finalizando alguns detalhes importantes

+++ src/com/compiladores/lexer/Real.java
@@ -1,9 +1,9 @@
 package com.compiladores.lexer;
 
 public class Real extends Token{
-    public final float value;
+    public final double value;
 
-    public Real(float v){
+    public Real(double v){
         super(Tag.REAL_CONST);
         this.value = v;
     }
